pyproject.toml
[tool.poetry]
name = "home-work-33"
version = "0.1.0"
description = ""
authors = ["Stas Kholshchevnikov <xcd85@mail.ru>"]
readme = "README.md"
packages = [{include = "home_work_33"}]

[tool.poetry.dependencies]
python = "^3.10"
envparse = "^0.2.0"
psycopg2-binary = "^2.9.6"
django = "^4.2.3"
djangorestframework = "^3.14.0"
social-auth-app-django = "^5.2.0"
django-filter = "^23.2"
pydantic = ">=1.10,<2.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
